% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{balance_dataset}
\alias{balance_dataset}
\title{Balance a Dataset for Model Training/Testing}
\usage{
balance_dataset(data, var)
}
\arguments{
\item{data}{A tibble with input data.}

\item{var}{The name of the column to balance.}
}
\value{

}
\description{
Classification models work best with "balanced" datasets where the number of
positive negative cases are roughly equal. This function takes an input
tibble and balances it by randomly downsampling until it has the same number
of positive and negative cases.
}
\details{
Note! It assumes there are more positive than negative cases because this is
always true in the customer-review datasets I'm working with. If your data
differs, the code is straightforward to modify.
}
