% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{flatten_stars}
\alias{flatten_stars}
\title{Flatten 5-Star Yelp Reviews to Positive/Negative}
\usage{
flatten_stars(data, var)
}
\arguments{
\item{data}{A tibble with Yelp reviews.}

\item{var}{The column containing star ratings as numeric values between 1 and
5.}
}
\value{
A modified tibble without 3-star ratings, and with a new column
containing "POS" for 4- and 5-star ratings and "NEG" for 1- and 2-star
ratings.
}
\description{
This function takes integer Yelp star ratings and "flattens" them into binary
positive/negative factor ratings. 1- and 2-star ratings become NEG, 4- and
5-star ratings become POS, and 3-star ratings are discarded.
}
