% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{get_accuracy}
\alias{get_accuracy}
\title{Get the accuracy of a set of Yelp review rating predictions.}
\usage{
get_accuracy(data, var)
}
\arguments{
\item{data}{A tibble of Yelp reviews that has been pre-processed with
\code{prepare_yelp()}, scored with \code{get_prob()}, and had a prediction made with
\code{predict_rating()}.}

\item{var}{The column containing true ratings.}
}
\value{
A one-value tibble with the total prediction accuracy.
}
\description{
Get the accuracy of a set of Yelp review rating predictions.
}
