% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{get_prob}
\alias{get_prob}
\title{Get the probability a Yelp review is positive.}
\usage{
get_prob(data, model_coefs = model_coeff_tibble)
}
\arguments{
\item{data}{A tibble of Yelp reviews pre-processed with \code{prepare_yelp()} that
has a column for mean AFINN score, the number of buts/nots, and each
review's word-length quintile.}

\item{model_coefs}{An optional tibble containing different model parameters.}
}
\value{
The original data with new columns for the log-odds and probability
that each review is positive.
}
\description{
Get the probability a Yelp review is positive.
}
