% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_coeff_tibble}
\alias{model_coeff_tibble}
\title{Model coefficients for predicting Yelp review ratings.}
\format{
A data frame with 5 rows and 3 variables.
}
\usage{
model_coeff_tibble
}
\description{
Coefficients for logistic functions to predict whether a Yelp review is positive or negative.
}
\keyword{datasets}
