% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_rating}
\alias{predict_rating}
\title{Predict whether a Yelp review is positive or negative.}
\usage{
predict_rating(data, threshold = 0.5)
}
\arguments{
\item{data}{A tibble of Yelp reviews pre-processed with \code{prepare_yelp()} and
scored with \code{get_prob()}.}

\item{threshold}{An optional threshold. Default is 0.5 which provided the
best accuracy in testing.}
}
\value{
The input data with a new column containing a predicted rating of
"POS" or "NEG" for each review.
}
\description{
Predict whether a Yelp review is positive or negative.
}
